Full Integration LLM/requirements.txt
streamlit
requests
python-dotenv
PyPDF2
pandas
langchain
langchain_google_genai
langchain_community
google-generativeai
scikit-learn
pytest